+++ mobile/android/android-components/components/browser/toolbar/src/test/java/mozilla/components/browser/toolbar/display/DisplayToolbarTest.kt
@@ -92,6 +92,7 @@ class DisplayToolbarTest {
         )
 
         assertNotNull(displayToolbar.views.trackingProtectionIndicator.colorFilter)
+        assertNotNull(displayToolbar.views.trackingProtectionIndicator.trackingProtectionTint)
     }
 
     @Test